Fix infinite recursion with prioritization in live mode
When presenting a prioritization content with live mode enabled, an
infinite recursion occurred related to the label cropping logic. While
Chrome and Firefox break the recursion and throw an error, on Safari the
browser tab becomes unresponsive when the infinite recursion occurs.

The root cause was that the calculation did not work as intended when
the chart was hidden.

This adjusts the recursion logic to insure it always is finite and
additionally adjusts the template so the stats component is not
initialized until it is needed.

Fixes: #1520