Fix submission of outdated/duplicate answers
There was an issue with event handling, which caused previous answers to
be submitted again with the current one. As a consequence, the backend
could register the wrong answer in some cases like answering the second
round or after the answers were reset. Futhermore, other related
requests like retrieving the correct answer were performed multiple
times after the submission.

An `EventEmitter` was used as an input to inform the child component
about the answer submission. This is a problem because they should only
be used for component outputs. For outputs, Angular automatically
handles unsubscription.

A `Subject` is now used and unsubscription is handled manually. The
naming has been adjusted because `sendEvent` can be easily be confused
for an action.

Fixes: #1522