Fix configuration root providers
This fixes an issue with our custom ErrorHandler being overridden.
The issue was cased by Angular's BrowserModule and
BrowserAnimationsModule which set up an ErrorHandler. With the migration
to standalone components, the order of providers changed which caused
this issue.

Instead of importing the old Angular modules, the provide methods are
now used instead, which fixes this issue.

Closes: #1535