Add global hint for read-only mode
This adds handling for the new global read-only mode. Outside of rooms,
a warning is shown when the mode is active. Furthermore, UI elements for
management on the room overview page and the room creation button are
disabled using the new `DisabledIfReadonly` directive.

This intentionally does not disable every single UI element affected by
read-only mode as this would require a more invasive approach with
modifications to a lot of components. Instead, error handling has been
extended to handle the 403 status code.

Currently, a few components do not have proper HTTP error handling. This
is not specific to read-only mode and should be fixed separately.

Introduce new `GlobalHintsComponent` which displays a hint for read-only mode.

For this extend `HintComponent` to support hints with border which are disabled by default and can be enabled with `showBorder` input.

Co-authored-by: Lukas Weingarten <lw_git@particify.de>

Closes: #1550