Improve starting, stopping and pausing the Q&A
Since the current workflow for toggling readonly mode and using the same icons for starting/stopping the Q&A (at room overview) can lead to confusion, this has been optimized.

Room overview:
You can now pause or stop the Q&A if started. If paused, the Q&A can be started or stopped. If stopped, the Q&A can be started (no matter if paused before).

Q&A page:
Use the same logic and icons as for the room overview. You can start the Q&A is paused or stopped. If started, you can pause the Q&A.

For this slightly improve a11y for `warn-soft` color in light theme.