Add share overlay component which can be displayed in presentation mode
New `ShareOverlayComponent` is used in presentation mode to permanently display joining instructions next to feature presentation.

The share overlay can be toggled with the QR code button and replaces the link (if shortener) or room code at the left upper edge. The state is stored in session storage to remember preference for current session.

If you hover over the share overlay, buttons for closing and expanding (share route) appear.