Improve highlighted comments in comment list
Instead of changing the background color of the whole comment card, simply use a big border in secondary color to indicate comment is highlighted.

This improves accessibility of elements inside the card and lead to a cleaner UI while highlighting is still visible enough.

To ensure correct spaces and border, use spacing variable for global card padding.