Fix focus mode initialization and subscriptions
The initialization was unreliable because it depended on init calls from
components, which is a bad practice in general. It worked for
presentation mode but caused issues when used from extensions.
Furthermore, RxJS subscriptions were not handled correctly which lead to
subscription leaks.

The way the code was written made it hard to fix the bug without the
risk of introducing new issues. Therefore, the code has been refactored
to simplify the logic. More improvements could be done, but for now this
only refactors the code which needs to be adjusted to fix the bugs.

Closes: particify/dev/extras/arsnova-webclient-extensions#97
Related: #1595