Fix focus mode initialization (follow up)
The previous fix did not subscribe to the observables in some cases, so
the variable assignment in tap had no effect in this case.

Instead of setting non-reactive variables through tap, Signal are now
used, which fixes this issue. Furthermore, comparison with the
previous state to inform participants did no longer work after the
recent changes and has been fixed. Handling of WS reconnects has been
integrated into the Observable pipe for the focus mode state.

The delay for initial navigation has been removed. Its purpose is
undocumented but I assume that it was a workaround for a race condition.
It should no longer be needed after the refactoring.

Closes: #1595, particify/dev/extras/arsnova-webclient-extensions#97