Solve a challenge before requesting guest account
A PoW-based challange is solved using Altcha. The challege is retrieved
from the backend and the solution is sent back in exchange for a JWT.
This can be used for authentication for endpoints which don't require a
user account. For now, it is used for creating guest accounts.

Related: particify/dev/foss/arsnova-server#344