Use new authentication API
The new API with separate refresh and access tokens is now used. The
client does not need to care about the refresh token. In fact, it cannot
even access it, since it is marked as HttpOnly for security reasons. It
still needs to call the refresh endpoint to trigger updating the cookie
and store the received access token, which it uses for further
authentication.

Because the client cannot delete the cookie, it now uses the new /logout
endpoint to let the backend handle it.

Related: particify/dev/foss/arsnova-server!988