Adjust to new GraphQL User type
The authentication provider has been removed. To keep the "guest"
account behavior, the new verified flag is used instead. Furthermore,
`displayId` and `displayName` are used instead of `loginId`. This allows
the display of user info in the navigation drawer once again.

The `AuthenticatedUser` now matches the GraphQL type. For now, a
separate type is still used to limit the needed refactoring but should
consider cleaning this up in the future.