Introduce new resolver for room ID
With GraphQL, having a resolver which fetches the whole room makes less
sense. The components should decide, which properties they need, and use
their own specialized queries. A resolver still makes sense to retrieve
the roomId for a shortId.

Furthermore, the resolver previously converted the roomId UUID to use
the legacy format without dashes. Components using GraphQL then had to
convert it back.

Changing this behavior across the whole application would require a lot
of refactoring. Therefore, the old resolver is kept and has been renamed
to `legacyRoomResolver`. It can be removed once the whole application
uses the GraphQL API. Therefore, a new `roomIdResolver` has been added.
This resolver only returns the roomId and should be used for new code
which uses the GraphQL API.

The `legacyRoomResolver` and `Room` are now deprecated.