Show logo splash screen while app is loading the first time
Add a static splash screen with minimalistic logo to `index.html` to display content immediately on first load.
This not only bridges the gap while the challenge is being solved, but also while the JS files are loading beforehand.

The result is a First Contentful Paint of a few milliseconds instead of several seconds.

As soon as Angular is ready, the splash screen disappears.
When reloading, no splash screen is necessary due to the service worker cache.

Related https://gitlab.com/particify/dev/foss/arsnova-webclient/-/issues/1612