Fix authentication interceptor
Requests to the token refresh endpoint should not be intercepted. The
access token does not need to be added since the cookie is added.
Previously, the interceptor could cause an endless request loop if the
token refresh endpoint responded with 401.

Other authentication endpoints do not need to be excluded explicitly
because they override the token and thereby disable interception logic.

Closes: #1634